+++ csp_helpers/templatetags/render_bundle_csp.py
@@ -1,7 +1,7 @@
-import contextlib
-
 from django import template
 from django.utils.safestring import mark_safe
+
+from csp_helpers.utils import get_nonce_from_context
 
 try:
     import webpack_loader
@@ -12,8 +12,9 @@ register = template.Library()
 
 
 def render_bundle_csp(context, bundle_name, extension=None, config='DEFAULT', attrs=''):
-    with contextlib.suppress(AttributeError):
-        attrs = ' '.join([attrs, f'nonce="{context.request.csp_nonce}"'])
+    nonce = get_nonce_from_context(context)
+    if nonce is not None:
+        attrs = ' '.join([attrs, f'nonce="{nonce}"'])
 
     tags = webpack_loader.utils.get_as_tags(bundle_name, extension=extension, config=config, attrs=attrs)
     return mark_safe('\n'.join(tags))